fix warnings

+++ es-core/src/components/ComponentList.cpp
@@ -179,7 +179,7 @@ void ComponentList::render(const Eigen::Affine3f& parentTrans)
 	for(unsigned int i = 0; i < mEntries.size(); i++)
 	{
 		auto& entry = mEntries.at(i);
-		drawAll = !mFocused || i != mCursor;
+		drawAll = !mFocused || i != (unsigned int)mCursor;
 		for(auto it = entry.data.elements.begin(); it != entry.data.elements.end(); it++)
 		{
 			if(drawAll || it->invert_when_selected)
